Remove `top_n_*`, `description_*`, `column_description`, `csv_column_extra` tables from PostgresQL
