--- src/column_statistics/save.rs
@@ -1,8 +1,0 @@
-mod binary;
-mod datetime;
-mod r#enum;
-mod float_range;
-mod int;
-mod ipaddr;
-pub(super) mod statistics;
-mod text;
