+++ src/migration.rs
@@ -2,7 +2,6 @@
 #![allow(clippy::too_many_lines)]
 
 mod migrate_classifiers_to_filesystem;
-mod migrate_column_stats;
 mod migrate_time_series;
 mod migration_structures;
 
@@ -142,7 +141,6 @@ pub async fn migrate_backend<P: AsRef<Path>>(
 
     migrate_classifiers_to_filesystem::run_migration(database).await?;
     if compatible.matches(&version) {
-        migrate_column_stats::run(database, store).await?;
         migrate_time_series::run(database, store).await?;
     }
